requirements.txt
fastapi
uvicorn[standard]
pydantic
websockets
ollama
python-multipart
gunicorn
python-dotenv
redis